+++ package.json
@@ -2,7 +2,8 @@
   "name": "swissqrbill",
   "version": "1.3.1",
   "description": "Swiss QR Bill generation in node.js ",
-  "main": "./lib/index.js",
+  "main": "./lib/node.js",
+  "browser": "./lib/browser.js",
   "types": "./lib/index.d.ts",
   "files": [
     "lib/**/*",
@@ -33,6 +34,7 @@
   },
   "homepage": "https://github.com/Rogerrrrrrrs/SwissQRBill#readme",
   "devDependencies": {
+    "@types/blob-stream": "^0.1.30",
     "@types/iban": "0.0.30",
     "@types/node": "^14.0.0",
     "@types/pdfkit": "^0.10.5",
@@ -42,9 +44,12 @@
     "@typescript-eslint/eslint-plugin": "^2.24.0",
     "@typescript-eslint/parser": "^2.24.0",
     "eslint": "^6.8.0",
-    "typescript": "^3.8.3"
+    "typescript": "^3.8.3",
+    "webpack": "^4.43.0",
+    "webpack-cli": "^3.3.12"
   },
   "dependencies": {
+    "blob-stream": "^0.1.3",
     "iban": "0.0.14",
     "pdfkit": "^0.11.0",
     "qrcode-svg": "^1.1.0",
